src/tasks/streamlit_app/requirements.txt
git+https://github.com/mokandil/pytube.git
edge_tts
ffmpeg
pydub
moviepy
translate
openai-whisper
streamlit
streamlit-option-menu
